Move tests for `Range` into new `RangeSpec` module.
Tests are now located in the same package as the code being tested.
